[processor/k8sattributes] Move packages to internal
